+++ nbo/predict.py
@@ -54,7 +54,7 @@ class DeploymentInfo:
 
 
 def _get_deployment_info(deployment_id: str) -> DeploymentInfo:
-    deployment = dr.Deployment.get(deployment_id)  # type: ignore[attr-defined]
+    deployment = dr.Deployment.get(deployment_id)
     target_name = deployment.model["target_name"]  # type: ignore[index]
     return DeploymentInfo(deployment, str(target_name))
 
@@ -116,7 +116,7 @@ def make_generative_deployment_predictions(
 
     for llm_request in requests:
         response = openai_client.chat.completions.create(
-            model="llm-blueprint",
+            model="datarobot-deployed-llm",
             messages=[
                 {"role": "system", "content": llm_request.system_prompt},
                 {"role": "user", "content": llm_request.prompt},